+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Variable name normalization with a sane default
 - `NormalizeFunc` to make variable normalization configurable
 
+### Changed
+
+- Global `Parse` does not return an error anymore
+
 
 ## [0.2.0] - 2018-06-06